requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
django-filter==23.5
drf-yasg==1.21.7
mysqlclient==2.2.1
requests==2.31.0
gunicorn==21.2.0
